PLANET-4603 Register campaign fields instead of metaboxes
* Use `register_post_meta` to add campaign fields instead of cmb2. The
interface for these fields will be added to
planet4-plugin-gutenberg-blocks.
* Use a wrapper function `P4_Post_Campaign::campaign_field()` to add
these fields, returning the field definition so that all fields can be
returned by a new REST API endpoint.
* Create a list of CSS variables. Ultimately these should be used as
actual css variables, however as a first step use these in a twig
template that only adds the rule if a variable is set. This intermediary
step is because a bigger change is needed to handle fallbacks when a
value is not set: the fallback would be dependant on which campaign
template is chosen.
* Simplify the logo setting so that it's not possible to choose the logo
of a different campaign. Only greenpeace/campaign.