PLANET-4603 Register campaign fields instead of metaboxes
* Use `register_post_meta` to add campaign fields instead of cmb2. The
editor interface for these fields will be added to
planet4-plugin-gutenberg-blocks.
* Move allowed values to json files. These are fetched by the frontend
to populate the values of the editor. Initially all files contain the
same values as were available for all templates in the code.
* Use a wrapper function `P4_Post_Campaign::campaign_field()` to add
these fields for convenience, as they all share some args.
* Create a list of CSS variables. Ultimately these should be used as
actual css variables, however as a first step use these in a twig
template that only adds the rule if a variable is set. This intermediary
step is because a bigger change is needed to handle fallbacks when a
value is not set: the fallback would be dependant on which campaign
template is chosen.